+++ packages/modules/cra-template-web-viewer/template/src/App.tsx
@@ -12,7 +12,7 @@ import {
   ScreenViewport,
   StandardViewId,
 } from "@itwin/core-frontend";
-import { Viewer } from "@itwin/web-viewer-react";
+import { Viewer } from "@itwin/web-viewer-react-3.0";
 import React, { useEffect, useState } from "react";
 
 import { Header } from "./Header";