+++ packages/modules/viewer-react/src/components/BaseBlankViewer.tsx
@@ -5,7 +5,7 @@
 
 import { BlankConnectionProps } from "@itwin/core-frontend";
 import { FillCentered } from "@itwin/core-react";
-import { ErrorBoundary } from "@itwin/error-handling-react";
+import { ErrorBoundary } from "@itwin/error-handling-react-3.0";
 import React, { useEffect, useState } from "react";
 
 import { BaseInitializer } from "..";