+++ packages/modules/web-viewer-react/src/index.ts
@@ -3,7 +3,7 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-export * from "@itwin/viewer-react";
+export * from "@itwin/viewer-react-3.0";
 export * from "./components/Viewer";
 export * from "./components/BlankViewer";
 export * from "./services/ItwinViewer";