+++ packages/modules/viewer-react/src/components/app-ui/frontstages/DefaultFrontstage.tsx
@@ -7,6 +7,7 @@ import { StageUsage, StandardContentLayouts } from "@itwin/appui-abstract";
 import {
   ContentGroup,
   CoreTools,
+  FrameworkVersion,
   Frontstage,
   FrontstageProvider,
   IModelViewportControl,
@@ -14,6 +15,8 @@ import {
   UiFramework,
   Widget,
   Zone,
+  ZoneLocation,
+  ZoneState,
 } from "@itwin/appui-react";
 import { ViewState } from "@itwin/core-frontend";
 import * as React from "react";
@@ -75,15 +78,18 @@ export class DefaultFrontstage extends FrontstageProvider {
   private _contentGroup: ContentGroup;
 
   private _uiConfig?: ItwinViewerUi;
+  private _frameworkVersion?: FrameworkVersion;
 
   constructor(
     public viewState: ViewState,
     uiConfig?: ItwinViewerUi,
-    viewportOptions?: ViewerViewportControlOptions
+    viewportOptions?: ViewerViewportControlOptions,
+    frameworkVersion?: FrameworkVersion
   ) {
     super();
 
     this._uiConfig = uiConfig;
+    this._frameworkVersion = frameworkVersion || "2";
 
     // Create the content group.
     this._contentGroup = new ContentGroup({
@@ -108,10 +114,11 @@ export class DefaultFrontstage extends FrontstageProvider {
     return (
       <Frontstage
         id="DefaultFrontstage"
+        version={1} // this value should be increased when changes are made to Frontstage
+        usage={StageUsage.General}
         defaultTool={CoreTools.selectElementCommand}
         contentGroup={this._contentGroup}
         isInFooterMode={true}
-        usage={StageUsage.General}
         contentManipulationTools={
           <Zone
             widgets={[
@@ -155,6 +162,19 @@ export class DefaultFrontstage extends FrontstageProvider {
               />,
             ]}
           />
+        }
+        centerRight={
+          this._frameworkVersion === "1" ? (
+            <Zone allowsMerging={true} defaultState={ZoneState.Minimized} />
+          ) : undefined
+        }
+        bottomRight={
+          this._frameworkVersion === "1" ? (
+            <Zone
+              allowsMerging={true}
+              mergeWithZone={ZoneLocation.CenterRight}
+            />
+          ) : undefined
         }
         statusBar={
           <Zone